Update 11ty and support streaming DSD
