+++ package.json
@@ -102,7 +102,7 @@
       ]
     },
     "serve": {
-      "command": "wds --watch --root-dir _prod",
+      "command": "npm run build:prod && wds --watch --root-dir _prod",
       "service": true,
       "dependencies": [
         {
@@ -132,20 +132,20 @@
   "dependencies": {
     "@11ty/is-land": "^3.0.0",
     "@material/mwc-button": "^0.27.0",
-    "lit": "^2.4.1"
+    "lit": "^2.7.0"
   },
   "devDependencies": {
-    "@11ty/eleventy": "^1.0.2",
-    "@lit-labs/eleventy-plugin-lit": "^0.2.3",
+    "@11ty/eleventy": "^2.0.1",
+    "@lit-labs/eleventy-plugin-lit": "^1.0.0",
     "@luncheon/esbuild-plugin-gzip": "^0.1.0",
     "@web/dev-server": "^0.1.35",
-    "@webcomponents/template-shadowroot": "^0.1.0",
+    "@webcomponents/template-shadowroot": "^0.2.1",
     "clean-css": "^5.3.1",
-    "esbuild": "^0.15.13",
+    "esbuild": "^0.17.14",
     "esbuild-plugin-minify-html-literals": "^1.0.1",
     "html-minifier": "^4.0.0",
     "tiny-glob": "^0.2.9",
-    "typescript": "^4.8.4",
-    "wireit": "^0.7.3"
+    "typescript": "^5.0.2",
+    "wireit": "^0.9.5"
   }
 }
